pyproject.toml
[build-system]
requires = [
  "setuptools",
  "setuptools_scm[toml]",
  "wheel",
]
build-backend = 'setuptools.build_meta'


[project]
name = "sunkit_pyvista"
dynamic = ["version"]
description = "3D visualisation of solar physics data with pyvista."
readme = "README.rst"
requires-python = ">=3.10"
license = {file = "LICENSE.txt"}
keywords = ["solar physics", "solar", "science", "3d", "plotting", "visualisation"]
authors = [
  { name = "The SunPy Community", email = "sunpy@googlegroups.com" }
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: BSD License",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Visualization",
]
dependencies = [
  'pyvista[all]>= 0.43.9',
  'sunpy[map]>=5.0.0',
]

[project.urls]
changelog = "https://docs.sunpy.org/projects/sunkit-magex/en/stable/changelog"
documentation = "https://docs.sunpy.org/projects/sunkit-magex/en/stable/"
homepage = "https://sunpy.org/"
issue_tracker = "https://github.com/sunpy/sunkit-pyvista/issues"
repository = "https://github.com/sunpy/sunkit-pyvista"

[project.optional-dependencies]
all = ["sunkit_pyvista"]
tests = [
  "sunkit_pyvista[all]",
  'pytest-astropy',
  "scikit-image",
  "sunkit-magex",
]
docs = [
  "sunkit_pyvista[all]",
  'sphinx',
  'sphinx-automodapi',
  'sphinx-changelog',
  'sphinx-gallery',
  'sunpy-sphinx-theme',
  'sunkit-magex',
  'sphinx-design',
]
dev = ["sunkit_pyvista[tests,docs]"]

[tool.setuptools_scm]
write_to = "sunkit_pyvista/_version.py"

[tool.setuptools]
include-package-data = true
platforms = ["any"]
provides = ["sunkit_pyvista"]
license-files = ["LICENSE.rst"]

[tool.setuptools.packages.find]
exclude = ["sunkit_pyvista._dev"]
namespaces = false

[ tool.gilesbot ]
  [ tool.gilesbot.pull_requests ]
    enabled = true

  [ tool.gilesbot.towncrier_changelog ]
    enabled = true
    verify_pr_number = true
    changelog_skip_label = "No Changelog Entry Needed"
    help_url = "https://github.com/sunpy/sunpy/blob/main/changelog/README.rst"
    changelog_missing_long = "There isn't a changelog file in this pull request. Please add a changelog file to the `changelog/` directory following the instructions in the changelog [README](https://github.com/sunpy/sunpy/blob/main/changelog/README.rst)."
    type_incorrect_long = "The changelog file you added is not one of the allowed types. Please use one of the types described in the changelog [README](https://github.com/sunpy/sunpy/blob/main/changelog/README.rst)"
    number_incorrect_long = "The number in the changelog file you added does not match the number of this pull request. Please rename the file."

[tool.towncrier]
    package = "sunkit_pyvista"
    filename = "CHANGELOG.rst"
    directory = "changelog/"
    issue_format = "`#{issue} <https://github.com/sunpy/sunkit-pyvista/pull/{issue}>`__"
    title_format = "{version} ({project_date})"

    [[tool.towncrier.type]]
      directory = "breaking"
      name = "Breaking Changes"
      showcontent = true

    [[tool.towncrier.type]]
      directory = "removal"
      name = "Removals"
      showcontent = true

    [[tool.towncrier.type]]
      directory = "deprecations"
      name = "Deprecations"
      showcontent = true

    [[tool.towncrier.type]]
      directory = "feature"
      name = "Features"
      showcontent = true

    [[tool.towncrier.type]]
      directory = "bugfix"
      name = "Bug Fixes"
      showcontent = true

    [[tool.towncrier.type]]
      directory = "doc"
      name = "Documentation"
      showcontent = true

    [[tool.towncrier.type]]
      directory = "trivial"
      name = "Internal Changes"
      showcontent = true
